+++ components/Footer.js
@@ -1,0 +1,15 @@
+const Footer = () => {
+  return (
+    <footer className='footer' aria-label='Footer'>
+      <div className='container'>
+        <div className='columns is-vcentered'>
+          <div className='column'>
+            <h1>I'm the Footer</h1>
+          </div>
+        </div>
+      </div>
+    </footer>
+  );
+};
+
+export default Footer;