+++ pages/_app.jsx
@@ -1,6 +1,9 @@
-import '../styles/globals.css'
+import 'bulma/css/bulma.min.css';
+import '../styles/globals.css';
+import MainLayout from '../layouts/MainLayout';
 
 const MyApp = ({ Component, pageProps }) => {
   return <Component {...pageProps} />
-}
-export default MyApp
+};
+
+export default MainLayout(MyApp);