+++ CHANGELOG.md
@@ -1,4 +1,11 @@
+PR Add styling and layout - 10/9/2021
+
+- Add Bulma packages
+- Update globals.css
+- Add Header and Footer components
+- Add main layout and wrap myApp with it
+
 PR Remove unnecessary files and packages - 10/8/2021
 
--   Remove unnecessary files and packages
--   Add favicon
+- Remove unnecessary files and packages
+- Add favicon